% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SimTable.phylo}
\alias{SimTable.phylo}
\title{Convert SimTable object to Phylogenetic Tree}
\usage{
SimTable.phylo(Robj, drop_extinct)
}
\arguments{
\item{Robj}{A tagged external pointer to a \code{sim_table_t} object, which contains the SimTable simulation results.}

\item{drop_extinct}{A logical flag indicating whether extinct lineages should be removed from the resulting phylogenetic tree.
\code{TRUE} will drop extinct lineages, and \code{FALSE} will return the full tree including extinct lineages.}
}
\value{
A phylo object (from the \code{ape} package) representing the phylogenetic tree extracted from the simulation table.
}
\description{
This function extracts a phylogenetic tree from a given SimTable object and returns it in a
phylo-compatible format, suitable for downstream analysis in R. It can optionally remove extinct lineages
from the tree.
}
\examples{
# Example usage
pars = c(0.5, 0.1, -0.001, -0.001, 0.0, 0.0)
sim_table <- evesim::edd_sim(pars = pars, age = 20, metric = "nnd", offset = "none")
phylo_tree <- SimTable.phylo(sim_table$sim, drop_extinct = TRUE)

}
\seealso{
\code{\link[ape]{phylo}} for more information about phylogenetic trees in R.
}
